requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Azure OpenAI
openai==1.59.5

# Database
mysql-connector-python==9.1.0
aiomysql==0.2.0
SQLAlchemy==2.0.36

# SQLite for thread storage
aiosqlite==0.20.0

# Data handling
pydantic==2.10.4
pydantic-settings==2.7.1
pandas==2.2.3

# CORS
python-dotenv==1.0.1

# Security
sqlparse==0.5.3
